Create principal_view.php
You will have the option of writing a query to perform a search or a logout button to close the session. For the graphical part we have implemented two search engines, one for the tables and one for the filters, this way it is easier to write the query, since we save the question mark.