Create filtres.php
It is called by pressing the send button, it will do the search in the database and will return to the main page the result in json where it will be decoded and displayed with a table on the screen. It is basically the code of the server.php adapted to the web page.